ofono: rework powerSet implementation to set Powered/Online properties always together
When the Powered property of a modem is set to true it's RF functionality is still turned
off. The user of the telephony service API expects the device to be in full functional
mode after issuing powerSet with it's parameter set to true. Therefore we have to set
Powered/Online always together.

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <morphis@gravedo.de>