Don't check for service state in every API method call
When an API method is called the serivce ensures that the underlaying service layer is
fully initialized. If the backend isn't initialized yet it's part of the corresponding
plugin to report this through the error object structure.

This is really important as otherwise subscriptions by clients in early lifetime of the
daemon aren't processed and get lost.

Open-WebOS-DCO-1.0-Signed-Off-By: Simon Busch <morphis@gravedo.de>